requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
werkzeug>=2.0.0
google-generativeai>=0.3.0
PyMuPDF>=1.22.0
python-docx>=0.8.11
python-pptx>=0.6.21
pytesseract>=0.3.10
pillow>=10.0.0
itsdangerous>=2.0.0
torch 
faiss-cpu 
transformers 
sentence-transformers
sentencepiece
sqlite3
